🔵 refacto: extract runMongoMap()
